bot/requirements.txt
python-telegram-bot==21.10
gspread==6.0.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
